% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_app.R
\name{cancer_map_usa}
\alias{cancer_map_usa}
\title{simple function for site-specific map}
\usage{
cancer_map_usa(site = "Stomach", scaling = 1)
}
\arguments{
\item{site}{character(1) defaults to "Stomach", must lie in value of woncan_types()}

\item{scaling}{numeric(1) scales the radius of circles, defaults to 1}
}
\description{
simple function for site-specific map
}
